Move testdata to internal directory
