+++ internal/blocktx/peer_handler_test.go
@@ -5,12 +5,11 @@ import (
 	"context"
 	"encoding/hex"
 	"errors"
+	"github.com/bitcoin-sv/arc/internal/testdata"
 	"log/slog"
 	"os"
 	"testing"
 	"time"
-
-	"github.com/bitcoin-sv/arc/testdata"
 
 	"github.com/bitcoin-sv/arc/internal/blocktx/blocktx_api"
 	"github.com/bitcoin-sv/arc/internal/blocktx/store"
