+++ internal/metamorph/processor_response_map_test.go
@@ -1,13 +1,13 @@
 package metamorph_test
 
 import (
+	"github.com/bitcoin-sv/arc/internal/testdata"
 	"testing"
 	"time"
 
 	"github.com/bitcoin-sv/arc/internal/metamorph"
 	"github.com/bitcoin-sv/arc/internal/metamorph/metamorph_api"
 	"github.com/bitcoin-sv/arc/internal/metamorph/processor_response"
-	"github.com/bitcoin-sv/arc/testdata"
 	"github.com/libsv/go-p2p/chaincfg/chainhash"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
